Change tests package name
