+++ src/test/java/delight/nashornsandbox/TestAllowAccess.java
@@ -1,8 +1,9 @@
-package delight.nashornsandbox.tests;
+package delight.nashornsandbox;
+
+import java.io.File;
 
 import javax.script.ScriptException;
 
-import org.junit.Assert;
 import org.junit.Test;
 
 import delight.nashornsandbox.NashornSandbox;
@@ -10,11 +11,11 @@ import delight.nashornsandbox.NashornSandboxes;
 import delight.nashornsandbox.exceptions.ScriptCPUAbuseException;
 
 @SuppressWarnings("all")
-public class TestSimpleEval {
+public class TestAllowAccess {
   @Test
-  public void test() throws ScriptCPUAbuseException, ScriptException {
+  public void test_file() throws ScriptCPUAbuseException, ScriptException {
     final NashornSandbox sandbox = NashornSandboxes.create();
-    final Object res = sandbox.eval("var x = 1 + 1; x;");
-    Assert.assertEquals(Integer.valueOf(2), res);
+    sandbox.allow(File.class);
+    sandbox.eval("var File = Java.type(\"java.io.File\"); File;");
   }
 }
