+++ src/test/java/delight/nashornsandbox/TestLimitCPU.java
@@ -1,4 +1,4 @@
-package delight.nashornsandbox.tests;
+package delight.nashornsandbox;
 
 import java.util.concurrent.Executors;
