+++ src/test/java/delight/nashornsandbox/TestManyEvalsAndInjections.java
@@ -1,4 +1,4 @@
-package delight.nashornsandbox.tests;
+package delight.nashornsandbox;
 
 import java.util.concurrent.Executors;
