+++ src/test/java/delight/nashornsandbox/TestSetWriter.java
@@ -1,4 +1,4 @@
-package delight.nashornsandbox.tests;
+package delight.nashornsandbox;
 
 import java.io.StringWriter;
